refactor: split generation helpers
